sagemaker/requirements.txt
pandas==2.1.1
numpy==1.24.3
boto3==1.28.36
tqdm==4.66.1
textblob==0.17.1
matplotlib==3.8.0
plotly==5.17.0
scikit-learn==1.3.1
wordcloud==1.9.2
seaborn==0.13.0
python-dotenv==1.0.0
pycountry==23.12.11
networkx==3.1
spacy==3.7.2